+++ packages/amplify-graphql-conversation-transformer/src/resolvers/invoke-lambda-resolver-fn.template.js
@@ -2,9 +2,10 @@ import { util } from '@aws-appsync/utils';
 
 export function request(ctx) {
   const { args, request } = ctx;
+  const { graphqlApiEndpoint } = ctx.stash;
+
   [[TOOL_DEFINITIONS_LINE]]
   const selectionSet = '[[SELECTION_SET]]';
-  const graphqlApiEndpoint = '[[GRAPHQL_API_ENDPOINT]]';
 
   const responseMutation = {
     name: '[[RESPONSE_MUTATION_NAME]]',