+++ packages/amplify-data-construct/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.11.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.10.2...@aws-amplify/data-construct@1.11.0) (2024-10-10)
+
+### Features
+
+- **conversation:** per message items and lambda history retrieval pattern ([#2914](https://github.com/aws-amplify/amplify-category-api/issues/2914)) ([874a30a](https://github.com/aws-amplify/amplify-category-api/commit/874a30ace18885f63c6be3274f32e4331bca98ed))
+
+## [1.10.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.10.1...@aws-amplify/data-construct@1.10.2) (2024-10-01)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
 ## [1.10.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.10.0...@aws-amplify/data-construct@1.10.1) (2024-09-16)
 
 **Note:** Version bump only for package @aws-amplify/data-construct