+++ packages/amplify-graphql-http-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@3.0.4...@aws-amplify/graphql-http-transformer@3.0.5) (2024-10-10)
+
+**Note:** Version bump only for package @aws-amplify/graphql-http-transformer
+
+## [3.0.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@3.0.3...@aws-amplify/graphql-http-transformer@3.0.4) (2024-10-01)
+
+**Note:** Version bump only for package @aws-amplify/graphql-http-transformer
+
 ## [3.0.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@3.0.2...@aws-amplify/graphql-http-transformer@3.0.3) (2024-09-16)
 
 **Note:** Version bump only for package @aws-amplify/graphql-http-transformer