+++ packages/amplify-graphql-transformer/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer@2.1.2...@aws-amplify/graphql-transformer@2.1.3) (2024-10-10)
+
+### Bug Fixes
+
+- **conversation:** use functionMap for custom handler IFunction reference ([#2922](https://github.com/aws-amplify/amplify-category-api/issues/2922)) ([d8d9eef](https://github.com/aws-amplify/amplify-category-api/commit/d8d9eefedc5ac411c9dc358a62c2461cfbb6a98a))
+
+## [2.1.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer@2.1.1...@aws-amplify/graphql-transformer@2.1.2) (2024-10-01)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer
+
 ## [2.1.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer@2.1.0...@aws-amplify/graphql-transformer@2.1.1) (2024-09-16)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer